+++ cli/src/admiral/client/netClient.go
@@ -1,3 +1,14 @@
+/*
+ * Copyright (c) 2016 VMware, Inc. All Rights Reserved.
+ *
+ * This product is licensed to you under the Apache License, Version 2.0 (the "License").
+ * You may not use this product except in compliance with the License.
+ *
+ * This product may include a number of subcomponents with separate copyright notices
+ * and license terms. Your use of these subcomponents is subject to the terms and
+ * conditions of the subcomponent's license, as noted in the LICENSE file.
+ */
+
 package client
 
 import (
@@ -11,6 +22,7 @@ import (
 	"time"
 
 	"admiral/auth"
+	"admiral/config"
 	"admiral/functions"
 )
 
@@ -18,16 +30,18 @@ type ResponseError struct {
 	Message string `json:"message"`
 }
 
+var timeoutDuration = time.Second * time.Duration(config.CLIENT_TIMEOUT)
+
 var netTransport = &http.Transport{
 	Dial: (&net.Dialer{
-		Timeout: 10 * time.Second,
+		Timeout: timeoutDuration,
 	}).Dial,
-	TLSHandshakeTimeout: 10 * time.Second,
+	TLSHandshakeTimeout: timeoutDuration,
 	Proxy:               http.ProxyFromEnvironment,
 }
 
 var NetClient = &http.Client{
-	Timeout:   time.Second * 10,
+	Timeout:   timeoutDuration,
 	Transport: netTransport,
 }