+++ cli/src/admiral/cmd/search.go
@@ -1,3 +1,14 @@
+/*
+ * Copyright (c) 2016 VMware, Inc. All Rights Reserved.
+ *
+ * This product is licensed to you under the Apache License, Version 2.0 (the "License").
+ * You may not use this product except in compliance with the License.
+ *
+ * This product may include a number of subcomponents with separate copyright notices
+ * and license terms. Your use of these subcomponents is subject to the terms and
+ * conditions of the subcomponent's license, as noted in the LICENSE file.
+ */
+
 package cmd
 
 import (
@@ -15,8 +26,8 @@ func init() {
 
 var searchCmd = &cobra.Command{
 	Use:   "search [IMAGE-NAME]",
-	Short: "Search for image.",
-	Long:  "Search for image.",
+	Short: "Search for image from which you can provision container.",
+	Long:  "Search for image from which you can provision container.",
 
 	Run: func(cmd *cobra.Command, args []string) {
 		if len(args) < 1 {