+++ cli/src/admiral/credentials/credentials.go
@@ -1,3 +1,14 @@
+/*
+ * Copyright (c) 2016 VMware, Inc. All Rights Reserved.
+ *
+ * This product is licensed to you under the Apache License, Version 2.0 (the "License").
+ * You may not use this product except in compliance with the License.
+ *
+ * This product may include a number of subcomponents with separate copyright notices
+ * and license terms. Your use of these subcomponents is subject to the terms and
+ * conditions of the subcomponent's license, as noted in the LICENSE file.
+ */
+
 package credentials
 
 import (
@@ -12,12 +23,12 @@ import (
 	"strings"
 	"syscall"
 
-	"golang.org/x/crypto/ssh/terminal"
-
 	"admiral/client"
 	"admiral/config"
 	"admiral/functions"
 	"admiral/properties"
+
+	"golang.org/x/crypto/ssh/terminal"
 )
 
 var (