+++ cli/src/admiral/functions/common.go
@@ -1,3 +1,14 @@
+/*
+ * Copyright (c) 2016 VMware, Inc. All Rights Reserved.
+ *
+ * This product is licensed to you under the Apache License, Version 2.0 (the "License").
+ * You may not use this product except in compliance with the License.
+ *
+ * This product may include a number of subcomponents with separate copyright notices
+ * and license terms. Your use of these subcomponents is subject to the terms and
+ * conditions of the subcomponent's license, as noted in the LICENSE file.
+ */
+
 package functions
 
 import (
@@ -32,4 +43,13 @@ func ValidateArgsCount(args []string) (string, bool) {
 		return args[0], true
 	}
 	return "", false
+}
+
+//ShortString is trimming string to desired length and adding three dots in the end of it.
+//In case len(s) <= desired length, the function returns the same string unmodified.
+func ShortString(s string, outputLen int) string {
+	if len(s) <= outputLen {
+		return s
+	}
+	return s[:outputLen-3] + "..."
 }