+++ cli/src/admiral/images/images.go
@@ -1,3 +1,14 @@
+/*
+ * Copyright (c) 2016 VMware, Inc. All Rights Reserved.
+ *
+ * This product is licensed to you under the Apache License, Version 2.0 (the "License").
+ * You may not use this product except in compliance with the License.
+ *
+ * This product may include a number of subcomponents with separate copyright notices
+ * and license terms. Your use of these subcomponents is subject to the terms and
+ * conditions of the subcomponent's license, as noted in the LICENSE file.
+ */
+
 package images
 
 import (
@@ -49,11 +60,7 @@ func (li *ImagesList) Print() {
 				trusted   string
 			)
 
-			if len(image.Description) > 40 {
-				desc = image.Description[:40] + "..."
-			} else {
-				desc = image.Description
-			}
+			desc = functions.ShortString(image.Description, 40)
 
 			if image.IsAutomated {
 				automated = "[OK]"