+++ cli/src/admiral/track/tracking.go
@@ -1,7 +1,19 @@
+/*
+ * Copyright (c) 2016 VMware, Inc. All Rights Reserved.
+ *
+ * This product is licensed to you under the Apache License, Version 2.0 (the "License").
+ * You may not use this product except in compliance with the License.
+ *
+ * This product may include a number of subcomponents with separate copyright notices
+ * and license terms. Your use of these subcomponents is subject to the terms and
+ * conditions of the subcomponent's license, as noted in the LICENSE file.
+ */
+
 package track
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -9,7 +21,6 @@ import (
 	"strings"
 	"time"
 
-	"errors"
 	"admiral/auth"
 	"admiral/client"
 	"admiral/config"
@@ -78,9 +89,7 @@ func Wait(taskId string) ([]string, error) {
 	begin := time.Now()
 	for {
 		elapsed := time.Now().Sub(begin)
-		if elapsed.Seconds() > float64(config.TIMEOUT) {
-			//fmt.Println("Task timed out.")
-			//os.Exit(0)
+		if elapsed.Seconds() > float64(config.TASK_TIMEOUT) {
 			return nil, errors.New("Task timed out.")
 		}
 		taskStatus := &TaskStatus{}
@@ -94,8 +103,6 @@ func Wait(taskId string) ([]string, error) {
 			os.Exit(-1)
 		}
 		if resp.StatusCode != 200 {
-			//fmt.Println("Resource not found")
-			//os.Exit(-1)
 			return nil, errors.New("Resource not found.")
 		}
 		err = json.Unmarshal(respBody, taskStatus)