+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.5.1-SNAPSHOT
 
+* Added Admiral CLI, **a command line tool to manage and automate Admiral**. More information [here] (https://github.com/vmware/admiral/blob/master/cli/README.md).
+
 ## 0.5.0
 
 * Initial open source release.